app/ml/image_processor.py
"""
Image processing and feature extraction using deep learning
"""

import cv2
import numpy as np
from PIL import Image
import torch
import torchvision.transforms as transforms
from typing import Dict, List, Tuple, Any
import logging
import io

logger = logging.getLogger(__name__)

class ImageProcessor:
    """
    Image processing and feature extraction for artwork analysis
    """
    
    def __init__(self):
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        self.transform = transforms.Compose([
            transforms.Resize((224, 224)),
            transforms.ToTensor(),
            transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225])
        ])
        self._load_models()
    
    def _load_models(self):
        """Load pre-trained models for feature extraction"""
        try:
            # TODO: Load actual pre-trained models
            # For now, use placeholder
            self.feature_extractor = None
            self.style_classifier = None
            logger.info("Image processing models loaded")
        except Exception as e:
            logger.warning(f"Could not load image models: {str(e)}")
    
    async def analyze_image(self, image_data: bytes, filename: str) -> Dict[str, Any]:
        """
        Comprehensive image analysis
        """
        try:
            # Load image
            image = Image.open(io.BytesIO(image_data))
            if image.mode != 'RGB':
                image = image.convert('RGB')
            
            # Basic image info
            width, height = image.size
            file_size = len(image_data)
            
            # Extract features
            features = await self.extract_features(image)
            
            # Analyze colors
            dominant_colors = self._extract_dominant_colors(image)
            
            # Detect objects/style
            detected_objects = await self._detect_objects(image)
            style_prediction = await self._predict_style(image)
            
            return {
                "filename": filename,
                "file_size": file_size,
                "dimensions": {"width": width, "height": height},
                "dominant_colors": dominant_colors,
                "detected_objects": detected_objects,
                "style_prediction": style_prediction,
                "features": features.tolist() if isinstance(features, np.ndarray) else features,
                "processing_time": 0.5  # Placeholder
            }
            
        except Exception as e:
            logger.error(f"Error analyzing image: {str(e)}")
            raise
    
    async def extract_features(self, image: Image.Image) -> np.ndarray:
        """
        Extract deep learning features from image
        """
        try:
            # Convert to tensor
            image_tensor = self.transform(image).unsqueeze(0).to(self.device)
            
            # TODO: Use actual CNN feature extractor
            # For now, return dummy features
            features = np.random.rand(512)
            
            return features
            
        except Exception as e:
            logger.error(f"Error extracting features: {str(e)}")
            return np.zeros(512)
    
    def _extract_dominant_colors(self, image: Image.Image, num_colors: int = 5) -> List[str]:
        """
        Extract dominant colors from image
        """
        try:
            # Convert to numpy array
            img_array = np.array(image)
            
            # Reshape to list of pixels
            pixels = img_array.reshape(-1, 3)
            
            # Use k-means clustering to find dominant colors
            from sklearn.cluster import KMeans
            
            kmeans = KMeans(n_clusters=num_colors, random_state=42, n_init=10)
            kmeans.fit(pixels)
            
            # Convert colors to hex
            colors = []
            for color in kmeans.cluster_centers_:
                hex_color = "#{:02x}{:02x}{:02x}".format(
                    int(color[0]), int(color[1]), int(color[2])
                )
                colors.append(hex_color)
            
            return colors
            
        except Exception as e:
            logger.error(f"Error extracting dominant colors: {str(e)}")
            return ["#000000", "#ffffff"]  # Fallback colors
    
    async def _detect_objects(self, image: Image.Image) -> List[str]:
        """
        Detect objects in the image
        """
        try:
            # TODO: Implement object detection
            # For now, return placeholder objects
            placeholder_objects = [
                "character", "background", "illustration", "artwork"
            ]
            return placeholder_objects[:2]  # Return first 2
            
        except Exception as e:
            logger.error(f"Error detecting objects: {str(e)}")
            return []
    
    async def _predict_style(self, image: Image.Image) -> Dict[str, float]:
        """
        Predict artwork style
        """
        try:
            # TODO: Implement style classification
            # For now, return placeholder predictions
            styles = {
                "anime": 0.7,
                "realistic": 0.2,
                "cartoon": 0.1,
                "digital": 0.8,
                "traditional": 0.2
            }
            
            return styles
            
        except Exception as e:
            logger.error(f"Error predicting style: {str(e)}")
            return {}
    
    def preprocess_for_similarity(self, image: Image.Image) -> np.ndarray:
        """
        Preprocess image for similarity comparison
        """
        try:
            # Resize to standard size
            image = image.resize((224, 224))
            
            # Convert to array and normalize
            img_array = np.array(image) / 255.0
            
            # Extract color histogram
            hist_r = cv2.calcHist([np.array(image)[:,:,0]], [0], None, [256], [0, 256])
            hist_g = cv2.calcHist([np.array(image)[:,:,1]], [0], None, [256], [0, 256])
            hist_b = cv2.calcHist([np.array(image)[:,:,2]], [0], None, [256], [0, 256])
            
            # Combine histograms
            color_features = np.concatenate([hist_r.flatten(), hist_g.flatten(), hist_b.flatten()])
            
            return color_features
            
        except Exception as e:
            logger.error(f"Error preprocessing image: {str(e)}")
            return np.zeros(768)  # 256*3 for RGB histograms
    
    def calculate_image_similarity(
        self,
        image1_features: np.ndarray,
        image2_features: np.ndarray
    ) -> float:
        """
        Calculate similarity between two images
        """
        try:
            from sklearn.metrics.pairwise import cosine_similarity
            
            similarity = cosine_similarity([image1_features], [image2_features])[0][0]
            return float(similarity)
            
        except Exception as e:
            logger.error(f"Error calculating similarity: {str(e)}")
            return 0.0