requirements.txt
# Core Framework
fastapi[all]==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Machine Learning
torch==2.1.1
torchvision==0.16.1
tensorflow==2.15.0
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
transformers==4.36.2

# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.1
sqlite3

# HTTP & API
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Cloudflare
cloudflare==2.13.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0